+++ monoruby/src/compiler/jitgen/asmir/compile/constants.rs
